Began rearranging code--both pvr and SDL versions of the md2 example use the same game.c file.
Increased size of vertex buffers to keep program from crashing when too
many polygons are drawn in one shot.

Changed md2 drawing code to only send the polygon header once per model
rather than once per triangle. This should be faster and not fill up
the vertex buffer as quickly.

Slightly optimized polygon drawing routines to use the vertex order
lookup table directly rather than copying it.

The md2 models now use the death animation when the entity dies in the
example program.

Dreamcast code now applies lighting to the md2 models based on the light
value of the cell in which each vertex of the rotated model is currently in.
It could probably be better, but it looks halfway decent for now.